ipx: split out backend file [#2254]
